Moved DisplayPreset implementation to Preset and added CONTRIBUTING.md
