+++ CONTRIBUTING.md
@@ -1,0 +1,5 @@
+# Setting up for development
+- Install .NET Core.
+- Make sure you have a Discord bot: [https://dsharpplus.github.io/articles/preamble.html](https://dsharpplus.github.io/articles/preamble.html).
+- Copy the dummy config file in `config/config.json` and paste it in `$HOME/marabot/config/config.json`.
+- Fill in the values in your pasted config file.
